Move common test config to libbeat
