+++ x-pack/filebeat/input/cloudfoundry/input_integration_test.go
@@ -8,7 +8,6 @@
 package cloudfoundry
 
 import (
-	"os"
 	"testing"
 	"time"
 
@@ -18,10 +17,11 @@ import (
 	"github.com/elastic/beats/v7/filebeat/input"
 	"github.com/elastic/beats/v7/libbeat/beat"
 	"github.com/elastic/beats/v7/libbeat/common"
+	cftest "github.com/elastic/beats/v7/x-pack/libbeat/common/cloudfoundry/test"
 )
 
 func TestInput(t *testing.T) {
-	config := common.MustNewConfigFrom(GetConfig(t))
+	config := common.MustNewConfigFrom(cftest.GetConfigFromEnv(t))
 
 	events := make(chan beat.Event)
 	connector := channel.ConnectorFunc(func(*common.Config, beat.ClientConfig) (channel.Outleter, error) {
@@ -41,42 +41,6 @@ func TestInput(t *testing.T) {
 		t.Logf("Event received: %+v", e)
 	case <-time.After(10 * time.Second):
 		t.Fatal("timeout waiting for events")
-	}
-}
-
-func GetConfig(t *testing.T) map[string]interface{} {
-	t.Helper()
-
-	config := map[string]interface{}{
-		"api_address":   lookupEnv(t, "CLOUDFOUNDRY_API_ADDRESS"),
-		"client_id":     lookupEnv(t, "CLOUDFOUNDRY_CLIENT_ID"),
-		"client_secret": lookupEnv(t, "CLOUDFOUNDRY_CLIENT_SECRET"),
-
-		"ssl.verification_mode": "none",
-	}
-
-	optionalConfig(config, "uaa_address", "CLOUDFOUNDRY_UAA_ADDRESS")
-	optionalConfig(config, "rlp_address", "CLOUDFOUNDRY_RLP_ADDRESS")
-	optionalConfig(config, "doppler_address", "CLOUDFOUNDRY_DOPPLER_ADDRESS")
-
-	if t.Failed() {
-		t.FailNow()
-	}
-
-	return config
-}
-
-func lookupEnv(t *testing.T, name string) string {
-	value, ok := os.LookupEnv(name)
-	if !ok {
-		t.Errorf("Environment variable %s is not set", name)
-	}
-	return value
-}
-
-func optionalConfig(config map[string]interface{}, key string, envVar string) {
-	if value, ok := os.LookupEnv(envVar); ok {
-		config[key] = value
 	}
 }
