+++ x-pack/metricbeat/module/cloudfoundry/mtest/config.go
@@ -5,44 +5,17 @@
 package mtest
 
 import (
-	"os"
 	"testing"
+
+	cftest "github.com/elastic/beats/v7/x-pack/libbeat/common/cloudfoundry/test"
 )
 
 func GetConfig(t *testing.T, metricset string) map[string]interface{} {
 	t.Helper()
 
-	config := map[string]interface{}{
-		"module":        "cloudfoundry",
-		"metricsets":    []string{metricset},
-		"api_address":   lookupEnv(t, "CLOUDFOUNDRY_API_ADDRESS"),
-		"client_id":     lookupEnv(t, "CLOUDFOUNDRY_CLIENT_ID"),
-		"client_secret": lookupEnv(t, "CLOUDFOUNDRY_CLIENT_SECRET"),
-
-		"ssl.verification_mode": "none",
-	}
-
-	optionalConfig(config, "uaa_address", "CLOUDFOUNDRY_UAA_ADDRESS")
-	optionalConfig(config, "rlp_address", "CLOUDFOUNDRY_RLP_ADDRESS")
-	optionalConfig(config, "doppler_address", "CLOUDFOUNDRY_DOPPLER_ADDRESS")
-
-	if t.Failed() {
-		t.FailNow()
-	}
+	config := cftest.GetConfigFromEnv(t)
+	config["module"] = "cloudfoundry"
+	config["metricsets"] = []string{metricset}
 
 	return config
-}
-
-func lookupEnv(t *testing.T, name string) string {
-	value, ok := os.LookupEnv(name)
-	if !ok {
-		t.Errorf("Environment variable %s is not set", name)
-	}
-	return value
-}
-
-func optionalConfig(config map[string]interface{}, key string, envVar string) {
-	if value, ok := os.LookupEnv(envVar); ok {
-		config[key] = value
-	}
 }
